// frontend/src/types/player.ts
export interface Player {
    id?: number;
    name: string;
    age: number;
    currentClub: string;
    league: string;
    position: string;
    marketValue: number;
    speed: number;
    acceleration: number;
    passing: number;
    vision: number;
    ballControl: number;
    finishing: number;
    dribbling: number;
    stamina: number;
    strength: number;
    tackling: number;
    positioning: number;
    workRate: number;
    createdAt?: string;
}